validacion de formularios a nivel usuario + orden de comentarios + order index

+++ prog-2-project/controllers/artworksController.js
@@ -1,17 +1,18 @@
+const { localsName } = require('ejs');
 var db = require('../database/models');
 
 
 
 const controller = {
-creator: function(req, res) {
-    db.Artwork.findAll({
-        'where': {'email': req.params.email}
-    }).then(function (result) {
-        res.render('product', { artworks: result });
-    }).catch(function (error) {
-        res.send(error);
-    })
-},
+// creator: function(req, res) {
+//     db.Artwork.findAll({
+//         'where': {'email': req.params.email}
+//     }).then(function (result) {
+//         res.render('product', { artworks: result });
+//     }).catch(function (error) {
+//         res.send(error);
+//     })
+// },
 show: function(req, res) {
     db.Artwork.findByPk(req.params.id, //{ 
         { include: 
@@ -23,7 +24,7 @@ show: function(req, res) {
                                //order: [ [ 'created_at', 'DESC'] ]
                             }
             ],
-            //order: [ [Comment, 'created_at', 'DESC'] ]
+           order: [ ['comments', 'created_at', 'DESC']] 
         },
        
         )
@@ -36,18 +37,24 @@ show: function(req, res) {
        
 },
 add: function(req, res) {
-//   if (!req.session.user) { 
-//         // res.redirect('/login'); //aca pondria register y en register pondria: all ready have an acount? Login here y el link
-//   throw Error('Not authorized.')
-//   }
-   res.render('artworks-add');
+    if(!req.session.user) {
+        return res.render ('artworks-add', { error: 'No perrmission. You must be registered or logged to add artworks'})
+    }
+    res.render('artworks-add');
 
  },
 store: function(req, res) {
-   // res.send(req.file)
-        req.body.user_id = req.session.user.id;
+   
         if (req.file) req.body.image = (req.file.path).replace('public', '');
+        if(!req.body.image) {
+            return res.render ('artworks-add', { noImage: 'You must include an image'})
+        }
+        if(!req.body.date) {
+            return res.render ('artworks-add', { noDate: 'You must include a date'})
+        }
+        req.body.user_id = req.session.user.id;
         req.body.created_at = new Date(); 
+      
         db.Artwork.create(req.body)
             .then(function(){
                 res.redirect('/')
@@ -72,8 +79,11 @@ store: function(req, res) {
     //     })
 },
 delete: function(req, res) {
-    if (!req.session.user) {
-        throw Error('Not authorized.') //creo que ya es medio al pedo porque directamente no te aparece el boton de delete
+    console.log(req.body.user_id);
+    console.log(req.session.user.id);
+    if (!req.session.user || req.session.user.id !== req.body.user_id) {
+       // throw Error('You are not the owner of the artwork you are trying to delete.')
+        return res.redirect ('/artworks/' + req.params.id) //creo que ya es medio al pedo porque directamente no te aparece el boton de delete
     }
     db.Artwork.destroy({ where: { id: req.params.id } })
         .then(function() {
@@ -84,9 +94,17 @@ delete: function(req, res) {
         })
 },
 edit: function(req, res) {
+    // if (!req.session.user || req.session.user.id !== req.body.user_id) {
+    //      return res.render ('artworks-add', {error: 'You are not the owner of the artwork you are trying to edit.'}) //creo que ya es medio al pedo porque directamente no te aparece el boton de delete
+    //  }
+
+    // console.log(req.body.user_id);
+    // console.log(req.session.user.id);
+    // console.log(req.body)
    db.Artwork.findByPk(req.params.id)
         .then(function (artworks) {
             res.render('artworks-edit', { artworks });
+            
         })
         .catch(function (error) {
             res.send(error);
@@ -96,6 +114,11 @@ edit: function(req, res) {
 update: function(req, res) {
     if (req.file) req.body.image = (req.file.path).replace('public', '');
     req.body.updated_at = new Date();
+
+//    console.log(req.body.user_id);
+//     console.log(req.session.user.id);
+//     console.log(req.body)
+   if(req.body.user_id == req.session.user.id){ //NO ME TOMA REQ.BODY.USER_ID LPM
     db.Artwork.update(req.body, { where: { id: req.params.id } })
         .then(function(artworks) {
             res.redirect('/artworks/' + req.params.id)
@@ -103,6 +126,9 @@ update: function(req, res) {
         .catch(function(error) {
             res.send(error);
         })
+    }else{
+        return res.render ('artworks-add', {error: 'You are not the owner of the artwork you are trying to edit.'})
+    }
     },
 comment: function(req, res) {
    if (!req.session.user) {
